backend/requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
geopandas==1.0.1
joblib==1.4.2
matplotlib==3.9.2
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.14.1
xgboost==2.1.1
